reslice origin
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@2642 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ DCMView.m
@@ -5731,17 +5731,17 @@ BOOL lineIntersectsRect(NSPoint lineStarts, NSPoint lineEnds, NSRect rect)
     glDisable (TEXTRECTMODE); // done with texturing
 }
 
--(float) pixelSpacing
+-(double) pixelSpacing
 {
 	return [curDCM pixelSpacingX];
 }
 
-- (float) pixelSpacingX
+- (double) pixelSpacingX
 {
 	return [curDCM pixelSpacingX];
 }
 
-- (float) pixelSpacingY
+- (double) pixelSpacingY
 {
 	return [curDCM pixelSpacingY];
 }
